requirements.txt
PyQt5Designer